+++ openeogeotrellis/geopysparkdatacube.py
@@ -1758,6 +1758,7 @@ class GeopysparkDataCube(DriverDataCube):
         * PNG: 8-bit grayscale or RGB raster with the limitation that it only export bands at a single (random) date
         * NetCDF: raster, currently using h5NetCDF
         * JSON: the json serialization of the underlying xarray, with extra attributes such as value/coord dtypes, crs, nodata value
+        * ZARR: chunked, compressed, N-dimensional arrays (experimental)
 
         :return: STAC assets dictionary: https://github.com/radiantearth/stac-spec/blob/master/item-spec/item-spec.md#assets
         """
@@ -2239,6 +2240,15 @@ class GeopysparkDataCube(DriverDataCube):
 
             XarrayIO.to_json_file(array=result, path=filename)
 
+        elif format == "ZARR":
+            band_names = ["var"]
+            if self.metadata.has_band_dimension():
+                band_names = self.metadata.band_names
+            number_bands = len(band_names)
+
+            self._save_zarr_executers(max_level,filename,number_bands)
+
+
         else:
             raise OpenEOApiException(
                 message="Format {f!r} is not supported".format(f=format),
@@ -2499,6 +2509,10 @@ class GeopysparkDataCube(DriverDataCube):
         else:
             return jvm.org.openeo.geotrellis.geotiff.package.saveStitchedTileGrid(spatial_rdd.srdd.rdd(), path,
                                                                                   tile_grid, max_compression)
+
+    def _save_zarr_executers(self,spatial_rdd, path, number_bands):
+        jvm = get_jvm()
+        jvm.org.openeo.geotrellis.zarr.zarrWriter.saveZarr(spatial_rdd,path,number_bands)
 
     @callsite
     def ndvi(self, **kwargs) -> 'GeopysparkDataCube':